bug fix for alternative voice commands for context names
Previously, when we were using <dgndictation>, we would use the overrides
dictionary to map triggers to the right context, but this broke with the change
to using set_list. To fix this, we need to add any additional commands we want
to the set_list so they can be triggered.